requirements.txt
alembic==1.0.9
astroid==2.2.5
Click==7.0
dominate==2.3.5
#Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
Mako==1.0.9
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.3.1
python-dateutil==2.8.0
python-editor==1.0.4
rope==0.14.0
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.1
visitor==0.1.3
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
